feat: Header 사용자 정보 조회 추가
