+++ src/components/organisms/Header/Header.styled.tsx
@@ -26,23 +26,9 @@ const LeftNav = styled.div`
   gap: 4rem;
 `;
 
-const RightNav = styled.div`
-  display: flex;
-  align-items: center;
-`;
-
-const NavLogin = styled.div`
-  display: flex;
-  align-items: center;
-  gap: 0.7rem;
-  height: 2.6rem;
-`;
-
 const S = {
   Header,
   LeftNav,
-  RightNav,
-  NavLogin,
 };
 
 export default S;
