+++ src/components/organisms/Header/RightNav/RightNav.styled.ts
@@ -1,0 +1,22 @@
+'use client';
+
+import styled from '@emotion/styled';
+
+const RightNav = styled.div`
+  display: flex;
+  align-items: center;
+`;
+
+const NavLogin = styled.div`
+  display: flex;
+  align-items: center;
+  gap: 0.7rem;
+  height: 2.6rem;
+`;
+
+const S = {
+  RightNav,
+  NavLogin,
+};
+
+export default S;
